✨ *refact(bot): reorganizar e simplificar a estrutura do bot*
As alterações visam melhorar a organização e a legibilidade do código do bot, removendo dependências desnecessárias e introduzindo uma nova classe para gerenciar propriedades.

Files Changed:
- *common/bot.py*: removido importação de Controller e simplificada a estrutura do código.
- *crawjud/core/__init__.py*: removido importações não utilizadas e alterada a classe CrawJUD para herdar de PropertiesCrawJUD.
- *crawjud/core/_properties.py*: adição de nova classe PropertiesCrawJUD para gerenciar variáveis e propriedades do bot.
- *crawjud/core/master.py*: arquivo removido, pois não é mais necessário.
