+++ crawjud/core/__init__.py
@@ -13,7 +13,6 @@ import unicodedata
 from contextlib import suppress
 from datetime import datetime
 from difflib import SequenceMatcher
-from logging import Logger
 from os import listdir, path
 from pathlib import Path
 from time import perf_counter, sleep
@@ -24,29 +23,21 @@ from cryptography import x509
 from cryptography.hazmat.backends import default_backend
 from pandas import Timestamp
 from pytz import timezone
-from selenium.webdriver.remote.webdriver import WebDriver
-from selenium.webdriver.support.wait import WebDriverWait
 from werkzeug.utils import secure_filename
 
 from addons.logger import dict_config
-from addons.printlogs import PrintMessage
-from crawjud.addons.auth import AuthController
-from crawjud.addons.elements import ElementsBot
 from crawjud.addons.make_templates import MakeTemplates
-from crawjud.addons.search import SearchController
 from crawjud.core._dictionary import BotData
-from crawjud.exceptions import AuthenticationError
+from crawjud.core._properties import PropertiesCrawJUD
 from crawjud.exceptions.bot import ExecutionError, StartError
-from crawjud.types import StrPath, TypeData
-from crawjud.types.elements import type_elements
+from crawjud.types import TypeData
 from models.logs import MessageLog
-from webdriver import DriverBot
 
 if TYPE_CHECKING:
     from crawjud.core._dictionary import BotData
 
 
-class CrawJUD:
+class CrawJUD(PropertiesCrawJUD):
     """Classe de controle de variáveis CrawJUD."""
 
     def initialize(
@@ -93,26 +84,8 @@ class CrawJUD:
             # Define o InputFile
             self.input_file = Path(self.output_dir_path).resolve().joinpath(self.xlsx)
 
-            # Instancia o WebDriver
-            self.configure_webdriver()
-
-            if self.system.lower() != "pje":
-                # Instancia o elements
-                self.elements = ElementsBot.config(
-                    system=self.system,
-                    state_or_client=self.state_or_client,
-                    **self.config_bot,
-                ).bot_elements
-
-                # Autenticação com os sistemas
-                self.portal_authentication()
-
             # Criação de planilhas template
             self.make_templates()
-
-            # Configura o search_bot
-            if self.system.lower() != "pje":
-                self.configure_searchengine()
 
             self.print_msg(
                 "Núcleo configurado.", self.pid, 0, "success", self.status_log
@@ -122,176 +95,6 @@ class CrawJUD:
 
         except Exception as e:
             raise StartError(exception=e) from e
-
-    _row: int = 0
-    # Variáveis de dados/configuraçoes
-    bot_data: dict[str, str]
-    config_bot: dict[str, AnyStr]
-    planilha_sucesso: StrPath
-    # Variáveis de estado/posição/indice
-    pid: str
-    pos: int
-    _is_stoped: bool
-    start_time: float
-
-    # Variáveis de verificações
-    system: str
-    typebot: str
-    state_or_client: str = None
-    preferred_browser: str = "chrome"
-    total_rows: int
-
-    # Variáveis de autenticação/protocolo
-    username: str
-    password: str
-    senhatoken: str
-
-    # Classes Globais
-    elements: type_elements
-    driver: WebDriver
-    search: SearchController
-    wait: WebDriverWait
-    logger: Logger
-    prt: PrintMessage
-
-    # Variáveis de nome/caminho de arquivos/pastas
-    xlsx: str
-    input_file: StrPath
-    output_dir_path: StrPath
-    _cities_am: dict[str, str]
-    _search: SearchController = None
-    _data_bot: dict[str, str] = {}
-
-    @property
-    def max_rows(self) -> int:  # noqa: D102
-        return self.prt.total_rows
-
-    @max_rows.setter
-    def max_rows(self, new_value: int) -> None:
-        self.prt.total_rows = new_value
-
-    @property
-    def total_rows(self) -> int:  # noqa: D102
-        return self.prt.total_rows
-
-    @total_rows.setter
-    def total_rows(self, new_value: int) -> None:
-        self.prt.total_rows = new_value
-
-    @property
-    def is_stoped(self) -> bool:  # noqa: D102
-        return self._is_stoped
-
-    @is_stoped.setter
-    def is_stoped(self, new_value: bool) -> None:
-        self._is_stoped = new_value
-
-    @property
-    def bot_data(self) -> BotData:
-        """Property bot data."""
-        return self._data_bot
-
-    @bot_data.setter
-    def bot_data(self, new_data: BotData) -> None:
-        """Property bot data."""
-        self._data_bot = new_data
-
-    @property
-    def search_bot(self) -> SearchController:
-        """Property para o searchbot."""
-        return self._search
-
-    @search_bot.setter
-    def search_bot(self, instancia: SearchController) -> None:
-        """Define a instância do searchbot."""
-        self._search = instancia
-
-    @property
-    def row(self) -> int:  # noqa: D102
-        return self._row
-
-    @row.setter
-    def row(self, new_value: int) -> None:
-        """Define o valor da variável row."""
-        self._row = new_value
-
-    @property
-    def cities_amazonas(self) -> dict[str, str]:  # noqa: N802
-        """Return a dictionary categorizing Amazonas cities as 'Capital' or 'Interior'.
-
-        Returns:
-            dict[str, str]: City names with associated regional classification.
-
-        """
-        return self._cities_am
-
-    def configure_searchengine(self) -> None:
-        """Configura a instância do search engine."""
-        self.search_bot = SearchController.construct(
-            system=self.system,
-            typebot=self.name,
-            driver=self.driver,
-            wait=self.wait,
-            elements=self.elements,
-            bot_data=self.bot_data,
-            prt=self.prt,
-        )
-
-    def portal_authentication(self) -> None:
-        """Autenticação com os sistemas."""
-        self.print_msg(
-            "Autenticando no sistema",
-            row=0,
-            type_log="log",
-            status=self.status_log,
-        )
-        auth = AuthController.construct(
-            system=self.system,
-            username=self.username,
-            password=self.password,
-            driver=self.driver,
-            wait=self.wait,
-            elements=self.elements,
-            prt=self.prt,
-        )
-        is_logged = auth.auth()
-
-        if not is_logged:
-            self.print_msg(
-                "Erro ao autenticar no sistema, verifique as credenciais.",
-                row=0,
-                type_log="error",
-                status=self.status_log,
-            )
-            raise AuthenticationError(
-                "Erro ao autenticar no sistema, verifique as credenciais."
-            )
-
-        self.print_msg(
-            "Autenticação realizada com sucesso",
-            row=0,
-            type_log="success",
-            status=self.status_log,
-        )
-
-    def configure_webdriver(self) -> None:
-        """Instancia o WebDriver."""
-        self.print_msg(
-            "Inicializando webdriver",
-            row=0,
-            type_log="log",
-            status=self.status_log,
-        )
-        self.driver = DriverBot(
-            self.preferred_browser, execution_path=self.output_dir_path
-        )
-        self.wait = self.driver.wait
-        self.print_msg(
-            "Webdriver inicializado",
-            row=0,
-            type_log="success",
-            status=self.status_log,
-        )
 
     def configure_logger(self) -> None:
         """Configura o logger."""
@@ -461,14 +264,6 @@ class CrawJUD:
 
     def tratamento_erros(self, exc: Exception, last_message: str = None) -> None:
         """Tratamento de erros dos robôs."""
-        windows = []
-        with suppress(Exception):
-            windows = self.driver.window_handles
-
-        if len(windows) == 0:
-            self.configure_webdriver()
-            self.portal_authentication()
-
         err_message = "\n".join(traceback.format_exception_only(exc))
         message = f"Erro de Operação: {err_message}"
         self.print_msg(message=message, type_log="error", pid=self.pid, row=self.row)
