--- crawjud/core/master.py
@@ -1,3 +1,0 @@
-"""Módulo de controle de variáveis CrawJUD."""
-
-from __future__ import annotations
